requirements.txt
mpegdash
tqdm
requests
python-dotenv
protobuf
webvtt-py
pysrt
m3u8
colorama
yt-dlp
bitstring
unidecode
beautifulsoup4
lxml
coloredlogs
browser_cookie3
demoji
deepl
fastapi
uvicorn
python-multipart
jinja2
firecrawl-py>=0.0.1